+++ src/inspector/components/DescriptorItem/DescriptorItem.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import "./DescriptorItem.less";
-import { IDescriptor, TSelectDescriptorOperation, ITargetReference, TAllReferenceSubtypes } from "../../model/types";
+import { IDescriptor, TSelectDescriptorOperation, ITargetReference } from "../../model/types";
 import { IconLockLocked, IconPin } from "../../../shared/components/icons";
 import { TState } from "../FilterButton/FilterButton";
-import { getName } from "../../classes/GetName";
 
 export interface IDescriptorItemProps {
 	descriptor: IDescriptor
@@ -44,49 +43,6 @@ export class DescriptorItem extends React.Component<TDescriptorItem> {
 		this.props.onSelect(this.props.descriptor.id, operation);
 	}
 
-	private generateItemNameNew = (): string => {
-		if (this.props.descriptor.originalReference.type==="listener") {
-			return this.props.descriptor.calculatedReference._obj;
-		}
-		const parts = getName(this.props.descriptor.calculatedReference._target);
-		//parts.push(...subs.map(d => d.subType + ": " + d.content.value));
-		const names = parts.map(p => /*p.typeTitle +*/ ((p.value) ? (/*": "*/ p.value) : p.typeTitle));
-		return names.join(" / ");
-	}
-
-	private generateItemName = () => {
-		if (this.props.activeTargetReference === null) { return "null - Error";}
-		// eslint-disable-next-line prefer-const
-		let { descriptor: { originalReference }, activeTargetReference,filterBySelectedReferenceType } = this.props;
-
-		const titles: string[] = [originalReference.type];
-		let subs:TAllReferenceSubtypes[] = [];
-		
-		if (originalReference.type === activeTargetReference.type && filterBySelectedReferenceType !== "off") {
-			titles.unshift();
-
-			subs = originalReference.data.filter((d,index) => {
-				const last = index === activeTargetReference.data.length - 1;
-				if (last) {
-					return true;
-				}
-				const found = activeTargetReference.data.find((a) => a.subType === d.subType);
-
-				if (!found) {
-					return true;
-				}
-
-				return (/*(found.content.value === d.content.value) &&*/ (found.content.filterBy === "off"));
-			});
-
-		} else {
-			subs = originalReference.data;
-		}
-			
-		titles.push(...subs.map(d => d.subType + ": " + d.content.value));
-		return titles.join(" / ");		
-	}
-
 	private get autoSelected():boolean {
 		return this.props.autoSelected.includes(this.props.descriptor.id);
 	}
@@ -97,7 +53,7 @@ export class DescriptorItem extends React.Component<TDescriptorItem> {
 		const {descriptor:{locked,pinned} } = this.props;
 		return (
 			<div className={"wrap" + (descriptor.selected ? " selected" : "") + (this.autoSelected?" autoSelected":"")} onClick={this.select}>
-				<div className="name">{this.generateItemNameNew()}</div>
+				<div className="name">{descriptor.title}</div>
 				<div className="spread"></div>
 				{locked ? <div className="icon"><IconLockLocked/></div> : ""}
 				{pinned ? <div className="icon"><IconPin/></div>: ""}