+++ src/inspector/components/ListenerFilter/ListenerFilterContainer.tsx
@@ -1,8 +1,8 @@
 import { IRootState } from "../../../shared/store";
 import { MapDispatchToPropsFunction, connect } from "react-redux";
 import { IListenerFilterProps, IListenerFilterDispatch, ListenerFilter } from "./ListenerFilter";
-import { getInspectorSettings, getSelectedTargetReference, getActiveTargetReferenceListenerCategory } from "../../selectors/inspectorSelectors";
-import { setExcludeAction, setFilterTypeAction, setIncludeAction, setFilterStateAction } from "../../actions/inspectorActions";
+import { getInspectorSettings, getSelectedTargetReference, getActiveTargetReferenceListenerCategory, getActiveTargetReference } from "../../selectors/inspectorSelectors";
+import { setExcludeAction, setFilterTypeAction, setIncludeAction, setFilterStateAction, setTargetReferenceAction } from "../../actions/inspectorActions";
 import { IContentWrapper, TListenerCategoryReference } from "../../model/types";
 
 
@@ -12,6 +12,7 @@ const mapStateToProps = (state: IRootState): IListenerFilterProps => {
 		settings: getInspectorSettings(state),
 		selectedTargetReference: getSelectedTargetReference(state),		
 		activeTargetReferenceListenerCategory: getActiveTargetReferenceListenerCategory(state) as IContentWrapper<TListenerCategoryReference>,
+		activeTargetReference: getActiveTargetReference(state),
 	};
 };
 
@@ -21,6 +22,7 @@ const mapDispatchToProps: MapDispatchToPropsFunction<IListenerFilterDispatch, Re
 		setFilterEventsType: (type) => dispatch(setFilterTypeAction(type)),
 		setInclude: (arr) => dispatch(setIncludeAction(arr)),
 		onSetFilter: (type, subType, state) => dispatch(setFilterStateAction(type, subType, state)),
+		onSetTargetReference:(arg) => dispatch(setTargetReferenceAction(arg)),
 	};
 };